+++ objectified-ui/public/WHATS_NEW.md
@@ -60,6 +60,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Added examples to schema request nodes in paths
   - Added encoding options for multipart in schema property nodes
   - Added link parameters assignable to HATEOAS link definitions
+  - Added response header templates to the response properties form
   - Dragging a property to a variable chip in the header will apply that property's primitive settings to the bound variable
   - Dragging a class to a response schema allows the class to be copied or referenced
 - OpenAPI Support: